Added Sentiment charge with two series: the current tweet's sentiment and the average sentiment over the last 5 seconds. TweetPerMinute-chart now shows the tweet-rate of the last 5 seconds, rather than minute. Both update rates mentioned above depend on the rate with which the statistics are polled, as the charts are updated in that callback.

+++ src/main/java/org/twintiment/analysis/AnalysisStatistics.java
@@ -122,7 +122,7 @@ public class AnalysisStatistics {
 	 *            sentimen value up-to-date.
 	 */
 	private void updateSentiment(double sentiment) {
-		avgSentiment = avgSentiment * ((numTweets - 1) / numTweets) + sentiment/numTweets;
+		avgSentiment = avgSentiment * ((numTweets - 1) / (double)numTweets) + sentiment/numTweets;
 		avgSentiment = Math.round(avgSentiment*10000)/10000d; //round to fourth decimal place	
 	}
